Split Android cloudsync account androidTests
