+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/account/credentialRecovery/LocalCloudAccountRepositoryCredentialRecoveryEraseTest.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository.cloudsync.account
+package com.flashcardsopensourceapp.data.local.repository.cloudsync.account.credentialRecovery
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.flashcardsopensourceapp.data.local.bootstrap.localWorkspaceName
